requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
firebase-admin
bcrypt
google-genai
torch
